Add earnings calculator utility and refactor view models
